+++ src/commands/ext.ts
@@ -1,6 +1,7 @@
 import * as _ from "lodash";
 import * as clc from "cli-color";
 
+import { checkMinRequiredVersion } from "../checkMinRequiredVersion";
 import { Command } from "../command";
 import * as getProjectId from "../getProjectId";
 import { logPrefix } from "../extensions/extensionsHelper";
@@ -14,6 +15,7 @@ module.exports = new Command("ext")
   .description(
     "display information on how to use ext commands and extensions installed to your project"
   )
+  .before(checkMinRequiredVersion, "extMinVersion")
   .action(async (options: any) => {
     // Print out help info for all extensions commands.
     utils.logLabeledBullet(logPrefix, "list of extensions commands:");