+++ src/emulator/downloadableEmulators.ts
@@ -17,6 +17,7 @@ import * as fs from "fs-extra";
 import * as path from "path";
 import * as os from "os";
 import { EmulatorRegistry } from "./registry";
+import { previews } from "../previews";
 
 // tslint:disable-next-line
 const downloadEmulator = require("../emulator/download");
@@ -28,14 +29,14 @@ const CACHE_DIR =
 
 const DownloadDetails: { [s in DownloadableEmulators]: EmulatorDownloadDetails } = {
   database: {
-    downloadPath: path.join(CACHE_DIR, "firebase-database-emulator-v4.5.0.jar"),
-    version: "4.5.0",
+    downloadPath: path.join(CACHE_DIR, "firebase-database-emulator-v4.6.0.jar"),
+    version: "4.6.0",
     opts: {
       cacheDir: CACHE_DIR,
       remoteUrl:
-        "https://storage.googleapis.com/firebase-preview-drop/emulator/firebase-database-emulator-v4.5.0.jar",
-      expectedSize: 28311004,
-      expectedChecksum: "1723857023077462f4b807922b1342f2",
+        "https://storage.googleapis.com/firebase-preview-drop/emulator/firebase-database-emulator-v4.6.0.jar",
+      expectedSize: 28458498,
+      expectedChecksum: "2b061fa9ba9f34d0d534f12caecddf3d",
       namePrefix: "firebase-database-emulator",
     },
   },
@@ -51,19 +52,36 @@ const DownloadDetails: { [s in DownloadableEmulators]: EmulatorDownloadDetails }
       namePrefix: "cloud-firestore-emulator",
     },
   },
-  ui: {
-    version: "1.1.1",
-    downloadPath: path.join(CACHE_DIR, "ui-v1.1.1.zip"),
-    unzipDir: path.join(CACHE_DIR, "ui-v1.1.1"),
-    binaryPath: path.join(CACHE_DIR, "ui-v1.1.1", "server.bundle.js"),
-    opts: {
-      cacheDir: CACHE_DIR,
-      remoteUrl: "https://storage.googleapis.com/firebase-preview-drop/emulator/ui-v1.1.1.zip",
-      expectedSize: 3248195,
-      expectedChecksum: "098821e328ea98c2180d4d71f3a75381",
-      namePrefix: "ui",
-    },
-  },
+  ui: previews.authemulator
+    ? {
+        version: "1.2.0-SNAPSHOT",
+        downloadPath: path.join(CACHE_DIR, "ui-1.2.0-SNAPSHOT.zip"),
+        unzipDir: path.join(CACHE_DIR, "ui-v1.2.0-SNAPSHOT"),
+        binaryPath: path.join(CACHE_DIR, "ui-v1.2.0-SNAPSHOT", "server.bundle.js"),
+        opts: {
+          cacheDir: CACHE_DIR,
+          remoteUrl:
+            "https://storage.googleapis.com/firebase-preview-drop/emulator/ui-v1.2.0-SNAPSHOT.zip",
+          expectedSize: -1,
+          expectedChecksum: "",
+          namePrefix: "ui",
+          skipChecksumAndSize: true,
+          skipCache: true,
+        },
+      }
+    : {
+        version: "1.1.1",
+        downloadPath: path.join(CACHE_DIR, "ui-v1.1.1.zip"),
+        unzipDir: path.join(CACHE_DIR, "ui-v1.1.1"),
+        binaryPath: path.join(CACHE_DIR, "ui-v1.1.1", "server.bundle.js"),
+        opts: {
+          cacheDir: CACHE_DIR,
+          remoteUrl: "https://storage.googleapis.com/firebase-preview-drop/emulator/ui-v1.1.1.zip",
+          expectedSize: 3248195,
+          expectedChecksum: "098821e328ea98c2180d4d71f3a75381",
+          namePrefix: "ui",
+        },
+      },
   pubsub: {
     downloadPath: path.join(CACHE_DIR, "pubsub-emulator-0.1.0.zip"),
     version: "0.1.0",