+++ src/previews.ts
@@ -5,8 +5,8 @@ interface PreviewFlags {
   rtdbrules: boolean;
   ext: boolean;
   extdev: boolean;
-  hostingchannels: boolean;
   rtdbmanagement: boolean;
+  authemulator: boolean;
 }
 
 export const previews: PreviewFlags = Object.assign(
@@ -15,8 +15,8 @@ export const previews: PreviewFlags = Object.assign(
     rtdbrules: false,
     ext: false,
     extdev: false,
-    hostingchannels: false,
     rtdbmanagement: false,
+    authemulator: false,
   },
   configstore.get("previews")
 );